google-sheet-mcp/requirements.txt
# Google API 相关依赖
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# 文档处理相关依赖
openpyxl>=3.1.0
python-docx>=0.8.11
pandas>=2.0.0

# MCP 服务相关依赖
mcp-server>=0.1.0

# 工具依赖
requests>=2.28.0
python-dateutil>=2.8.2
tqdm>=4.65.0
